requirements.txt
torch
torchvision
einops
timm
flask
Pillow
Flask-HTTPAuth
werkzeug
opencv-python